+++ transit_odp/naptan/dataclasses/__init__.py
@@ -1,0 +1,9 @@
+from .naptan import StopPoint
+from .nptg import NationalPublicTransportGazetteer, NptgLocality, Region
+
+__all__ = [
+    "NationalPublicTransportGazetteer",
+    "NptgLocality",
+    "Region",
+    "StopPoint",
+]