+++ transit_odp/pipelines/pipelines/naptan_etl/extract.py
@@ -1,14 +1,18 @@
 import os
 import shutil
 import zipfile
+from pathlib import Path
 
 import pandas as pd
 import requests
 from celery.utils.log import get_task_logger
 from django.conf import settings
 from lxml import etree as ET
+from requests import RequestException
 
-logger = get_task_logger(__name__)
+from transit_odp.common.loggers import LoaderAdapter
+from transit_odp.naptan.dataclasses import StopPoint
+from transit_odp.naptan.dataclasses.nptg import NationalPublicTransportGazetteer
 
 ns = "http://www.naptan.org.uk/"
 namespace = {"naptan": ns}
@@ -22,46 +26,56 @@ DISK_PATH_FOR_NAPTAN_FOLDER = "/tmp/NaptanStops/"
 DISK_PATH_FOR_NPTG_FOLDER = "/tmp/NPTG/"
 
 
-def get_latest_naptan_xml():
-    logger.info("[get_latest_naptan_xml] called")
-    xml_file_path = None
-    try:
-        naptan_url = settings.NAPTAN_IMPORT_URL
-        response = requests.get(naptan_url)
-        response.encoding = "utf-8"
+logger = get_task_logger(__name__)
+logger = LoaderAdapter("NaPTANLoader", logger)
 
-        logger.info("Writing response data to a file on disk")
+
+def get_latest_naptan_xml():
+    naptan_url = settings.NAPTAN_IMPORT_URL
+    logger.info(f"Loading NaPTAN file from {naptan_url}.")
+    xml_file_path = None
+
+    try:
+        response = requests.get(naptan_url)
+    except RequestException as exc:
+        logger.error(f"Unable to fetch NaPTAN data from {naptan_url}.", exc_info=exc)
+        return xml_file_path
+
+    try:
+        logger.info("Writing NaPTAN response data to a file on disk.")
         if response.status_code == 200:
-            with open(DISK_PATH_FOR_NAPTAN_ZIP, "wb") as f:
+            dir_path = Path(DISK_PATH_FOR_NAPTAN_FOLDER)
+            if not dir_path.exists():
+                dir_path.mkdir(parents=True, exist_ok=True)
+
+            filepath = dir_path / "Naptan.xml"
+            with filepath.open("wb") as f:
                 for chunk in response.iter_content(CHUNK_SIZE):
                     f.write(chunk)
-            logger.info("Finished writing to file on disk")
 
-            # Now that we have the zip file on disk, extract zip
-            with zipfile.ZipFile(DISK_PATH_FOR_NAPTAN_ZIP, "r") as zip_ref:
-                zip_ref.extractall(DISK_PATH_FOR_NAPTAN_FOLDER)
-
+            logger.info("Finished NaPTAN writing to file on disk.")
             for filename in os.listdir(DISK_PATH_FOR_NAPTAN_FOLDER):
                 if filename.endswith(".xml"):
                     xml_file_path = os.path.join(DISK_PATH_FOR_NAPTAN_FOLDER, filename)
-    except Exception:
-        logger.info("Exception while getting Naptan data")
-    logger.info("File extract done")
+    except Exception as exc:
+        logger.error("Exception while getting NaPTAN data.", exc_info=exc)
+
+    logger.info("NaPTAN file successfully extracted to disk.")
     return xml_file_path
 
 
 def get_latest_nptg():
-    logger.info("[get_latest_nptg] called")
+    nptg_url = settings.NPTG_IMPORT_URL
+    logger.info(f"Loading NPTG from {nptg_url}.")
     xml_file_path = None
     try:
-        nptg_url = settings.NPTG_IMPORT_URL
         response = requests.get(nptg_url)
-        logger.info("Writing response data to a file on disk")
+        logger.info("Writing NPTG response data to a file on disk.")
         if response.status_code == 200:
             with open(DISK_PATH_FOR_NPTG_ZIP, "wb") as f:
                 for chunk in response.iter_content(CHUNK_SIZE):
                     f.write(chunk)
-            logger.info("Finished writing to file on disk")
+            logger.info("Finished NPTG writing to file on disk.")
 
             # Now that we have the zip file on disk, extract zip
             with zipfile.ZipFile(DISK_PATH_FOR_NPTG_ZIP, "r") as zip_ref:
@@ -71,37 +85,30 @@ def get_latest_nptg():
                 if filename.endswith(".xml"):
                     xml_file_path = os.path.join(DISK_PATH_FOR_NPTG_FOLDER, filename)
     except Exception:
-        logger.info("Exception while getting NPTG data")
+        logger.info("Exception while getting NPTG data.")
 
-    logger.info("File extract done")
+    logger.info("NPTG file successfully extracted to disk.")
     return xml_file_path
 
 
 def extract_stops(xml_file_path):
     def inner():
-        logger.info("[extract_stops] called")
+        logger.info(f"Extracting NaPTAN stops from file {xml_file_path}.")
         tree = ET.parse(xml_file_path)
-        root = tree.getroot()
 
-        for stop in root.iter(f"{{{ns}}}StopPoint"):
+        stop_point_path = "//naptan:StopPoints/naptan:StopPoint"
+        for stop in tree.iterfind(stop_point_path, namespaces=namespace):
+            point = StopPoint.from_xml(stop)
             yield {
-                "atco_code": stop.findtext("naptan:AtcoCode", namespaces=namespace),
-                "naptan_code": stop.findtext("naptan:NaptanCode", namespaces=namespace),
-                "common_name": stop.findtext(
-                    ".//naptan:CommonName", namespaces=namespace
-                ),
-                "indicator": stop.findtext(".//naptan:Indicator", namespaces=namespace),
-                "street": stop.findtext(".//naptan:Street", namespaces=namespace),
-                "locality_id": stop.findtext(
-                    ".//naptan:NptgLocalityRef", namespaces=namespace
-                ),
-                "admin_area_id": int(
-                    stop.findtext(
-                        ".//naptan:AdministrativeAreaRef", namespaces=namespace
-                    )
-                ),
-                "latitude": stop.findtext(".//naptan:Latitude", namespaces=namespace),
-                "longitude": stop.findtext(".//naptan:Longitude", namespaces=namespace),
+                "atco_code": point.atco_code,
+                "naptan_code": point.naptan_code,
+                "common_name": point.descriptor.common_name,
+                "indicator": point.descriptor.indicator,
+                "street": point.descriptor.street,
+                "locality_id": point.place.nptg_locality_ref,
+                "admin_area_id": int(point.administrative_area_ref),
+                "latitude": point.place.location.translation.latitude,
+                "longitude": point.place.location.translation.longitude,
             }
 
     df = pd.DataFrame(
@@ -118,6 +125,7 @@ def extract_stops(xml_file_path):
             "longitude",
         ],
     )
+    logger.info(f"A total of {len(df)} NaPTAN stops extracted.")
 
     duplicated = df["atco_code"].duplicated(keep="first")
     duplicates = df[duplicated]
@@ -129,91 +137,53 @@ def extract_stops(xml_file_path):
         df = df[~duplicated]
 
     df = df.set_index("atco_code", verify_integrity=True)
-    logger.info("[extract_stops] finished")
+
+    logger.info(f"After deduplication {len(df)} NaPTAN stops extracted.")
     return df
 
 
 def extract_admin_areas(xml_file_path):
     def inner():
-        logger.info("[extract_admin_areas] called")
+        logger.info(f"Extracting NPTG AdminAreas from {xml_file_path}.")
         tree = ET.parse(xml_file_path)
         root = tree.getroot()
+        nptg = NationalPublicTransportGazetteer.from_xml(root)
 
-        for region in root.iter(f"{{{ns}}}Region"):
-            for admin_area in region.iter(f"{{{ns}}}AdministrativeArea"):
+        for region in nptg.regions:
+            region_code = region.region_code
+            for area in region.administrative_areas:
                 yield {
-                    "id": int(
-                        admin_area.findtext(
-                            "naptan:AdministrativeAreaCode", namespaces=namespace
-                        )
-                    ),
-                    "name": admin_area.findtext("naptan:Name", namespaces=namespace),
-                    "traveline_region_id": region.findtext(
-                        "naptan:RegionCode", namespaces=namespace
-                    ),
-                    "atco_code": admin_area.findtext(
-                        "naptan:AtcoAreaCode", namespaces=namespace
-                    ),
+                    "id": int(area.administrative_area_code),
+                    "name": area.name,
+                    "traveline_region_id": region_code,
+                    "atco_code": area.atco_area_code,
                 }
 
     df = pd.DataFrame(
         inner(),
         columns=["id", "name", "traveline_region_id", "atco_code"],
     ).set_index("id", verify_integrity=True)
-    logger.info("[extract_admin_areas] finished")
-    return df
 
-
-def extract_districts(xml_file_path):
-    def inner():
-        logger.info("[extract_districts] called")
-        tree = ET.parse(xml_file_path)
-        root = tree.getroot()
-
-        for district in root.iter(f"{{{ns}}}NptgDistrict"):
-            yield {
-                "id": int(
-                    district.findtext("naptan:NptgDistrictCode", namespaces=namespace)
-                ),
-                "name": district.findtext("naptan:Name", namespaces=namespace),
-            }
-
-    df = pd.DataFrame(
-        inner(),
-        columns=["id", "name"],
-    ).set_index("id", verify_integrity=True)
-    logger.info("[extract_localities] finished")
+    logger.info(f"{len(df)} NPTG AdminAreas extracted.")
     return df
 
 
 def extract_localities(xml_file_path):
+    logger.info(f"Extracting NPTG Localities from {xml_file_path}.")
+
     def inner():
-        logger.info("[extract_localities] called")
         tree = ET.parse(xml_file_path)
         root = tree.getroot()
+        nptg = NationalPublicTransportGazetteer.from_xml(root)
 
-        for locality in root.iter(f"{{{ns}}}NptgLocality"):
+        for locality in nptg.nptg_localities:
             yield {
-                "gazetteer_id": locality.findtext(
-                    "naptan:NptgLocalityCode", namespaces=namespace
-                ),
-                "name": locality.findtext(
-                    ".//naptan:LocalityName", namespaces=namespace
-                ),
-                "easting": int(
-                    locality.findtext(".//naptan:Easting", namespaces=namespace)
-                ),
-                "northing": int(
-                    locality.findtext(".//naptan:Northing", namespaces=namespace)
-                ),
-                "district_id": int(
-                    locality.findtext("naptan:NptgDistrictRef", namespaces=namespace)
-                ),
-                "admin_area_id": int(
-                    locality.findtext(
-                        "naptan:AdministrativeAreaRef", namespaces=namespace
-                    )
-                ),
+                "gazetteer_id": locality.nptg_locality_code,
+                "name": locality.descriptor.locality_name,
+                "easting": locality.location.translation.easting,
+                "northing": locality.location.translation.northing,
+                "district_id": int(locality.nptg_district_ref),
+                "admin_area_id": int(locality.administrative_area_ref),
             }
 
     df = pd.DataFrame(
@@ -228,16 +198,21 @@ def extract_localities(xml_file_path):
         ],
     ).set_index("gazetteer_id", verify_integrity=True)
 
-    logger.info("[extract_localities] finished")
+    logger.info(f"{len(df)} NPTG Localities extracted.")
     return df
 
 
 def cleanup():
+    logger.info("Cleaning up files saved to disk.")
     if os.path.exists(DISK_PATH_FOR_NAPTAN_ZIP):
+        logger.info(f"Removing {DISK_PATH_FOR_NAPTAN_ZIP}.")
         os.remove(DISK_PATH_FOR_NAPTAN_ZIP)
     if os.path.exists(DISK_PATH_FOR_NAPTAN_FOLDER):
+        logger.info(f"Removing {DISK_PATH_FOR_NAPTAN_FOLDER}.")
         shutil.rmtree(DISK_PATH_FOR_NAPTAN_FOLDER)
     if os.path.exists(DISK_PATH_FOR_NPTG_ZIP):
+        logger.info(f"Removing {DISK_PATH_FOR_NPTG_ZIP}.")
         os.remove(DISK_PATH_FOR_NPTG_ZIP)
     if os.path.exists(DISK_PATH_FOR_NPTG_FOLDER):
+        logger.info(f"Removing {DISK_PATH_FOR_NPTG_FOLDER}.")
         shutil.rmtree(DISK_PATH_FOR_NPTG_FOLDER)